+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## [0.1.0] - Initial release
-* added **Seven-segment** and **Fourteen-segment** display
+## [0.2.0]
+* added **sixteen-segment** display
+* updated segment styles for sixteen-segment display
+* updated HexSegmentStyle diagonals
+
+## [0.1.0]
+* initial release
+* added **seven-segment** and **fourteen-segment** display
 * added DefaultSegmentStyle, HexSegmentStyle and RectSegmentStyle